Move the validation part of the BackgroundCommandWorker to a separate file.
